Develop

+++ composeApp/src/commonMain/kotlin/com/nnzapp/easymarket/domain/model/CartItem.kt
@@ -2,8 +2,8 @@ package com.nnzapp.easymarket.domain.model
 
 data class CartItem(
     val product: Product,
-    val quantity: Int
+    val quantity: Double,
 ) {
-    val totalPrice: Int
+    val totalPrice: Double
         get() = product.price * quantity
 }
